+++ main.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/hashicorp/vault/api"
 	"github.com/hashicorp/vault/sdk/plugin"
-	"github.com/omisego/immutability-eth-plugin/ethereum"
+	"github.com/omgnetwork/immutability-eth-plugin/ethereum"
 )
 
 func main() {